cart and checkout to api

+++ StyleHub/Models/product.cs
@@ -4,12 +4,12 @@
     {
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string? Size { get; set; }
         public int Review { get; set; }
         public bool InStock { get; set; }
         public string? Badge { get; set; }
-        public double? OldPrice { get; set; }
+        public decimal? OldPrice { get; set; }
         public double Rating { get; set; }
         public string? Color { get; set; }
         public List<string>? Images { get; set; } // لو الصور محفوظة كـ URL
